+++ spec/lib/pipedrive/resource_spec.rb
@@ -231,6 +231,18 @@ RSpec.describe Pipedrive::Resourceable do
         it "returns an Resourceable instance with no data" do
           expect(subject.empty?).to be_truthy
         end
+
+        context "treat 204 as 404" do
+          before do
+            allow(Pipedrive).to receive(:treat_no_content_as_not_found).and_return(true)
+          end
+
+          it "raises 404 Not found error" do
+            expect do
+              described_class.retrieve(1)
+            end.to raise_error Pipedrive::NotFoundError
+          end
+        end
       end
     end